% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_dataframes.R
\name{AvgBedgraph}
\alias{AvgBedgraph}
\title{Take a merged bedgraph dataframe, and return a dataframe with the mean average bedgraph coverage for each record}
\usage{
AvgBedgraph(df)
}
\arguments{
\item{df}{A merged bedgraph dataframe in which each record represent one locus on a give chromosome, and columns represent coverage values from different read mapping analyses}
}
\value{
df A bedgraph dataframe with the artihmetic mean average coverage for each records locus
}
\description{
Take a merged bedgraph dataframe, and return a dataframe with the mean average bedgraph coverage for each record
}
