% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_dataframes.R
\name{AvgSalmonQuant}
\alias{AvgSalmonQuant}
\title{Get the average abundance per records for multiple salmon runs}
\usage{
AvgSalmonQuant(df)
}
\arguments{
\item{df}{A merged salmon dataframe in which each record represent one locus on a give chromosome, and columns represent coverage values from different read quasi-mapping analyses}
}
\value{
df A bedgraph dataframe with the artihmetic mean average coverage for each records locus
}
\description{
Get the average abundance per records for multiple salmon runs
}
