% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{filter_contextpp_scores}
\alias{filter_contextpp_scores}
\title{Filter context++ scores by the normalised expression values of putative target transcripts within a given biological context}
\usage{
filter_contextpp_scores(contextpp_scores_filename,
  expression_values_filename)
}
\arguments{
\item{contextpp_scores_filename}{The file name of a context++ results file}

\item{expression_values_filename}{The file name of a salmon output file of normalised expression values}
}
\value{
A tibble of context++ scores filtered by a given TPM threshold
}
\description{
Filter context++ scores by the normalised expression values of putative target transcripts within a given biological context
}
