% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mirs.R
\name{filter_mature_mirs}
\alias{filter_mature_mirs}
\title{Filter full set of mature miRNAs by a list of user selected miRNAs}
\usage{
filter_mature_mirs(input_file, list_of_mirs, output_file)
}
\arguments{
\item{input_file}{A tab delimited file of mature miRNA sequences and associated identifiers}

\item{list_of_mirs}{A vector of miRNA names}

\item{output_file}{Name of file to write to}
}
\description{
Filter full set of mature miRNAs by a list of user selected miRNAs
}
