% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mirs.R
\name{filter_mir_families}
\alias{filter_mir_families}
\title{Filter miRNA family data according to a vector of miRNA name}
\usage{
filter_mir_families(mature_mirna_file, mirna_families_file, list_of_mirs,
  output_file)
}
\arguments{
\item{mature_mirna_file}{file name of mature miRNA data}

\item{mirna_families_file}{file name of miRNA family data}

\item{list_of_mirs}{A vector of miRNA names}

\item{output_file}{A filtered set of miRNA family data}
}
\description{
Filter miRNA family data according to a vector of miRNA name
}
