% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{filter_miRanda_scores}
\alias{filter_miRanda_scores}
\title{Filter miRanda scores by the normalised expression values of putative target transcripts within a given biological context}
\usage{
filter_miRanda_scores(miRanda_scores_filename, expression_values_filename)
}
\arguments{
\item{miRanda_scores_filename}{The file name of a miRanda results file}

\item{expression_values_filename}{The file name of a salmon output file of normalised expression values}
}
\value{
A tibble of miRanda scores filtered by a given TPM threshold
}
\description{
Filter miRanda scores by the normalised expression values of putative target transcripts within a given biological context
}
