% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{fix_ts_output}
\alias{fix_ts_output}
\title{fix output from a modified version of the targetscan script which finds seed matches}
\usage{
fix_ts_output(ts_sites_output)
}
\arguments{
\item{ts_sites_output}{- Output from the modified version of the first targetscan script}
}
\value{
A corrected targetscan seed match output file
}
\description{
fix output from a modified version of the targetscan script which finds seed matches
}
