% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{get_AIR_file}
\alias{get_AIR_file}
\title{Get a full AIR file specific to a given coding transcriptome}
\usage{
get_AIR_file(APA_file, UTR_lengths_file)
}
\arguments{
\item{APA_file}{- The output from APAtrap's APA prediction perl script}

\item{UTR_lengths_file}{- A dataframe containing transcript accessions with corresponding 3'UTR lengths}
}
\description{
Get a full AIR file specific to a given coding transcriptome
}
