% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIR_processing.R
\name{get_canonical_AIRs}
\alias{get_canonical_AIRs}
\title{Generate a file of affected isoform ratio (AIR) data for use in computing context++ scores when executing targetscan7 (10.7554/eLife.05005).}
\usage{
get_canonical_AIRs(utr_lengths)
}
\arguments{
\item{utr_lengths}{A two-column tab delimited file: tx_id, utr_length}
}
\value{
AIRs A four-column tibble - tx_id, start, utr_length, AIR
}
\description{
Generate a file of affected isoform ratio (AIR) data for use in computing context++ scores when executing targetscan7 (10.7554/eLife.05005).
}
