% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{get_mirna_context}
\alias{get_mirna_context}
\title{get miRNA file needed for computing contextpp scores}
\usage{
get_mirna_context(mirna_seed_file, mature_mirnas_file, species)
}
\arguments{
\item{mirna_seed_file}{A file containing miRNA accessions with corresponding seed sequences}

\item{mature_mirnas_file}{A file containing miRNA accessions with corresponding mature miRNA sequences}

\item{species}{- three-letter species code}
}
\value{
A file containing the name of the miRNA family, the NCBI species taxonomic ID, the name of the mature miRNA and also the mature miRNA sequence
}
\description{
get miRNA file needed for computing contextpp scores
}
