% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscan.R
\name{get_mirna_family}
\alias{get_mirna_family}
\title{Generates a miRNA family file in specific targetscan format}
\usage{
get_mirna_family(mirna_seeds, species)
}
\arguments{
\item{species}{three letter species code e.g. hsa,mmu etc - used as the reference species}

\item{mirna_seed}{information - a table with miRNA and species identifiers as well as seed sequences}
}
\value{
A data frame in which miRNA famiilies have been enumerated, with seed sequences and NCBI species taxnonomic IDs. All families contain a seed sequence from the reference species
}
\description{
Generates a miRNA family file in specific targetscan format
}
