% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utr_lengths.R
\name{get_utr_lengths}
\alias{get_utr_lengths}
\title{Returns the length of 3'UTRs from bed data}
\usage{
get_utr_lengths(utr_bed_file)
}
\arguments{
\item{utr_bed_file}{A bed 6 file of 3'UTRs}
}
\value{
A two-column dataframe of transcript IDs and corresponding 3'UTR lengths
}
\description{
Returns the length of 3'UTRs from bed data
}
