% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFasta.R
\name{MergeFasta}
\alias{MergeFasta}
\title{Concatenate fasta seqeunce for fasta records with the same header. Concatenation occurs in order of appearance in the file}
\usage{
MergeFasta(file)
}
\arguments{
\item{name}{of file containing fasta records}
}
\value{
A one-column tibble in which fasta headers and fasta sequences are interleaved
}
\description{
Concatenate fasta seqeunce for fasta records with the same header. Concatenation occurs in order of appearance in the file
}
